Fix voiceover mentioning the wrong li tag
Due to the code logic changing the all DOM lis, voiceover detected a change of a li which was actually not changed (because the js code was resetting all
li to false to put them back to true, so already checked li would appear like having been changed by voiceover, when the human
been eye could not see the change). The new code logic only modify the DOM of what really changed.